requirements.txt
absl-py==1.4.0
altair==5.1.2
annoy==1.17.3
asgiref==3.6.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.1
certifi==2021.10.8
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Django==4.2
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
image==1.5.33
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.3
streamlit==1.28.2
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
validators==0.22.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.17.0